break CloudWatch tests into appender-only and writer-only
