+++ appenders/src/test/resources/TestCloudWatchAppender/testStaticClientFactory.properties
@@ -1,4 +1,4 @@
-# config for mock-object tests of writer operation; note short batch delay
+# config for client factory test
 
 log4j.rootLogger=DEBUG, default
 
@@ -8,4 +8,4 @@ log4j.appender.default.layout=org.apache.log4j.PatternLayout
 log4j.appender.default.logGroup=argle
 log4j.appender.default.logStream=bargle
 
-log4j.appender.default.clientFactory=com.kdgregory.log4j.aws.TestCloudWatchAppender.createMockClient
+log4j.appender.default.clientFactory=com.kdgregory.log4j.aws.TestCloudWatchLogWriter.createMockClient
